feat: 배치 executor 설정
